code/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.0.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
exceptiongroup==1.2.0
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
flask-restful-swagger-3==0.5.1
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pathspec==0.12.1
pip-chill==1.0.3
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
